requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-restframework==0.0.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
kombu==5.1.0
prompt-toolkit==3.0.19
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
psycopg2-binary>=2.8
